requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
Flask==3.0.2
gunicorn==21.2.0
Werkzeug==3.0.1

# ── HTTP ───────────────────────────────────────────────────────────────────────
requests==2.31.0

# ── Data & ML ─────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==2.0.0
scikit-learn==1.8.0
joblib==1.4.2
openpyxl==3.1.2

# ── PDF Generation ─────────────────────────────────────────────────────────────
reportlab==4.0.8

# ── Environment ────────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ── Date/Time ─────────────────────────────────────────────────────────────────
python-dateutil==2.8.2
pytz==2024.1
